backend/pyproject.toml
[tool.poetry]
name = "agroclimavisio-backend"
version = "1.0.0"
description = "Backend FastAPI pour AgroClimaVisio"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
xarray = "^2024.1.0"
netcdf4 = "^1.6.4"
rasterio = "^1.3.9"
numpy = "^1.26.0"
geopandas = "^0.14.0"
shapely = "^2.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

